backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
celery[redis]==5.3.6
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
pytesseract==0.3.10
numpy==1.26.4
opencv-python-headless==4.9.0.80
pdfplumber==0.10.3
pdf2image==1.17.0
Pillow==10.2.0
vaderSentiment==3.3.2
transformers==4.38.1
spacy==3.7.2
ftfy==6.1.3
langdetect==1.0.9
boto3==1.34.34
python-magic==0.4.27
python-multipart==0.0.9
redis==5.0.1
websockets==12.0
fpdf2==2.7.8
reportlab==4.1.0
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
aiosqlite==0.20.0
cloudinary==1.40.0
